refactor: rename kubernetes-context-state-constants to contexts-constants Signed-off-by: Philippe Martin <phmartin@redhat.com>
